2 cameras, tree assets centered, new Basic Map, stretching sprite bugfix
After a lot of testing with making a sudo camera with a mask around the game area, I realized that you can have 2 cameras in one scene and so I switched to that. I kept the ground tracker I made for the sudo camera to move the new cameras.

The tree assets were not centered or cropped well so I went to every singe one of them and made them a multiple sprite and then sliced them with the auto function.

There was a bug that was stretching on the x axis every sprite we put inside the basic map. Like the actual backgrounds, obstacles, finish line and in the future anything else we would add. Although it took me literal ages to figure it out, I realized it was the way the play area was created. So I fixed this by making a new Basic Map, this time the right way.