my stuff + a few changes from jerry
Dash is also explained in "latest dash script" push, but: you can set the "ammo" for the dash in the PlayerController script, in the editor (no need to open the script). Any apple objects need a collider that is trigger and the tag "DashAmmo" (should be in the list of tags). I'm not sure I destroy the apple on collision right now, but it gives you +1 dash ammo. Dashing gives you -1 ammo.